moved MarathonTaskStatus to own proper package
